modif de outline de Jeu

+++ src/main/java/menuprincipal/battleship/joueur/Joueur.java
@@ -3,13 +3,13 @@ package menuprincipal.battleship.joueur;
 import javafx.util.Pair;
 
 public abstract class Joueur {
-    public Pair<Double, Double> determinerTir() {
-        Pair<Double, Double> tir = null;
+    public Pair<Integer, Integer> determinerTir() {
+        Pair<Integer, Integer> tir = null;
         return null;
     }
 
-    public Pair<Double, Double> demanderTir() {
-        Pair<Double, Double> tir = null;
+    public Pair<Integer, Integer> demanderTir() {
+        Pair<Integer, Integer> tir = null;
         return null;
     }
