+++ src/main/java/menuprincipal/battleship/plateau/Plateau.java
@@ -4,6 +4,6 @@ import javafx.util.Pair;
 
 public abstract class Plateau {
     protected Case[][] cases;
-    public void ajouterTir(Pair<Double, Double> coordonee) {
+    public void ajouterTir(Pair<Integer, Integer> coordonee) {
     }
 }
